IW: Don't include the Traceur runtime in dev
I was unable to get the Traceur runtime to be included into the build without
messing up the source map line numbers. This commit includes the runtime
in a different script tag for development. It also includes the runtime
in the compressed version of the tangram. Users of the compressed
version should not noticed a difference.